Move & rename metrics scripts
